requirements.txt
fastapi
uvicorn
requests
python-dotenv
wemake-python-styleguide==0.19.2
